+++ glN64_GX/N64.h
@@ -16,31 +16,31 @@
 
 struct N64Regs
 {
-	unsigned long *MI_INTR;
+	unsigned int *MI_INTR;
 
-	unsigned long *DPC_START;
-	unsigned long *DPC_END;
-	unsigned long *DPC_CURRENT;
-	unsigned long *DPC_STATUS;
-	unsigned long *DPC_CLOCK;
-	unsigned long *DPC_BUFBUSY;
-	unsigned long *DPC_PIPEBUSY;
-	unsigned long *DPC_TMEM;
+	unsigned int *DPC_START;
+	unsigned int *DPC_END;
+	unsigned int *DPC_CURRENT;
+	unsigned int *DPC_STATUS;
+	unsigned int *DPC_CLOCK;
+	unsigned int *DPC_BUFBUSY;
+	unsigned int *DPC_PIPEBUSY;
+	unsigned int *DPC_TMEM;
 
-	unsigned long *VI_STATUS;
-	unsigned long *VI_ORIGIN;
-	unsigned long *VI_WIDTH;
-	unsigned long *VI_INTR;
-	unsigned long *VI_V_CURRENT_LINE;
-	unsigned long *VI_TIMING;
-	unsigned long *VI_V_SYNC;
-	unsigned long *VI_H_SYNC;
-	unsigned long *VI_LEAP;
-	unsigned long *VI_H_START;
-	unsigned long *VI_V_START;
-	unsigned long *VI_V_BURST;
-	unsigned long *VI_X_SCALE;
-	unsigned long *VI_Y_SCALE;
+	unsigned int *VI_STATUS;
+	unsigned int *VI_ORIGIN;
+	unsigned int *VI_WIDTH;
+	unsigned int *VI_INTR;
+	unsigned int *VI_V_CURRENT_LINE;
+	unsigned int *VI_TIMING;
+	unsigned int *VI_V_SYNC;
+	unsigned int *VI_H_SYNC;
+	unsigned int *VI_LEAP;
+	unsigned int *VI_H_START;
+	unsigned int *VI_V_START;
+	unsigned int *VI_V_BURST;
+	unsigned int *VI_X_SCALE;
+	unsigned int *VI_Y_SCALE;
 };
 
 extern N64Regs REG;
